requirements.txt
nltk
librosa
numpy
scipy
python-dotenv
requests
pyphen
underthesea

# New dependencies for OCR script
pytesseract
Pillow
yt-dlp

paddleocr
paddlepaddle

openai-whisper